Update from upstream

+++ traffic_ops/traffic_ops_golang/role/roles.go
@@ -24,6 +24,7 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
+	"time"
 
 	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-tc"
@@ -44,21 +45,32 @@ type TORole struct {
 	PQCapabilities *pq.StringArray `json:"-" db:"capabilities"`
 }
 
+func (v *TORole) GetLastUpdated() (*time.Time, bool, error) {
+	return api.GetLastUpdated(v.APIInfo().Tx, *v.ID, "role")
+}
+
+func (v *TORole) SelectMaxLastUpdatedQuery(where, orderBy, pagination, tableName string) string {
+	return `SELECT max(t) from (
+		SELECT max(last_updated) as t from ` + tableName + ` r ` + where + orderBy + pagination +
+		` UNION ALL
+	select max(last_updated) as t from last_deleted l where l.table_name='` + tableName + `') as res`
+}
+
 func (v *TORole) SetLastUpdated(t tc.TimeNoMod) { v.LastUpdated = &t }
 func (v *TORole) InsertQuery() string           { return insertQuery() }
 func (v *TORole) NewReadObj() interface{}       { return &TORole{} }
 func (v *TORole) SelectQuery() string           { return selectQuery() }
 func (v *TORole) ParamColumns() map[string]dbhelpers.WhereColumnInfo {
 	return map[string]dbhelpers.WhereColumnInfo{
-		"name": dbhelpers.WhereColumnInfo{"name", nil},
-		"id":   dbhelpers.WhereColumnInfo{"id", api.IsInt},
-	}
+		"name":      dbhelpers.WhereColumnInfo{Column: "name"},
+		"id":        dbhelpers.WhereColumnInfo{Column: "id", Checker: api.IsInt},
+		"privLevel": dbhelpers.WhereColumnInfo{Column: "priv_level", Checker: api.IsInt}}
 }
 func (v *TORole) UpdateQuery() string { return updateQuery() }
 func (v *TORole) DeleteQuery() string { return deleteQuery() }
 
 func (role TORole) GetKeyFieldsInfo() []api.KeyFieldInfo {
-	return []api.KeyFieldInfo{{"id", api.GetIntKey}}
+	return []api.KeyFieldInfo{{Field: "id", Func: api.GetIntKey}}
 }
 
 //Implementation of the Identifier, Validator interface functions
@@ -122,9 +134,11 @@ func (role *TORole) Create() (error, error, int) {
 	}
 
 	//after we have role ID we can associate the capabilities:
-	userErr, sysErr, errCode = role.createRoleCapabilityAssociations(role.ReqInfo.Tx)
-	if userErr != nil || sysErr != nil {
-		return userErr, sysErr, errCode
+	if role.Capabilities != nil && len(*role.Capabilities) > 0 {
+		userErr, sysErr, errCode = role.createRoleCapabilityAssociations(role.ReqInfo.Tx)
+		if userErr != nil || sysErr != nil {
+			return userErr, sysErr, errCode
+		}
 	}
 	return nil, nil, http.StatusOK
 }
@@ -156,33 +170,50 @@ func (role *TORole) deleteRoleCapabilityAssociations(tx *sqlx.Tx) (error, error,
 	return nil, nil, http.StatusOK
 }
 
-func (role *TORole) Read() ([]interface{}, error, error, int) {
-	vals, userErr, sysErr, errCode := api.GenericRead(role)
-	if userErr != nil || sysErr != nil {
-		return nil, userErr, sysErr, errCode
+func (role *TORole) Read(h http.Header, useIMS bool) ([]interface{}, error, error, int, *time.Time) {
+	version := role.APIInfo().Version
+	api.DefaultSort(role.APIInfo(), "name")
+	vals, userErr, sysErr, errCode, maxTime := api.GenericRead(h, role, useIMS)
+	if errCode == http.StatusNotModified {
+		return []interface{}{}, nil, nil, errCode, maxTime
 	}
+	if userErr != nil || sysErr != nil {
+		return nil, userErr, sysErr, errCode, maxTime
+	}
+
+	returnable := []interface{}{}
 	for _, val := range vals {
 		rl := val.(*TORole)
-		caps := ([]string)(*rl.PQCapabilities)
-		rl.Capabilities = &caps
+		switch {
+		case version.Major > 1 || version.Minor >= 3:
+			caps := ([]string)(*rl.PQCapabilities)
+			rl.Capabilities = &caps
+			returnable = append(returnable, rl)
+		case version.Minor >= 1:
+			returnable = append(returnable, rl.RoleV11)
+		}
 	}
-	return vals, nil, nil, http.StatusOK
+	return returnable, nil, nil, http.StatusOK, maxTime
 }
 
-func (role *TORole) Update() (error, error, int) {
+func (role *TORole) Update(h http.Header) (error, error, int) {
 	if *role.PrivLevel > role.ReqInfo.User.PrivLevel {
 		return errors.New("can not create a role with a higher priv level than your own"), nil, http.StatusForbidden
 	}
-	userErr, sysErr, errCode := api.GenericUpdate(role)
+	userErr, sysErr, errCode := api.GenericUpdate(h, role)
 	if userErr != nil || sysErr != nil {
 		return userErr, sysErr, errCode
 	}
+
 	// TODO cascade delete, to automatically do this in SQL?
-	userErr, sysErr, errCode = role.deleteRoleCapabilityAssociations(role.ReqInfo.Tx)
-	if userErr != nil || sysErr != nil {
-		return userErr, sysErr, errCode
+	if role.Capabilities != nil && *role.Capabilities != nil {
+		userErr, sysErr, errCode = role.deleteRoleCapabilityAssociations(role.ReqInfo.Tx)
+		if userErr != nil || sysErr != nil {
+			return userErr, sysErr, errCode
+		}
+		return role.createRoleCapabilityAssociations(role.ReqInfo.Tx)
 	}
-	return role.createRoleCapabilityAssociations(role.ReqInfo.Tx)
+	return nil, nil, http.StatusOK
 }
 
 func (role *TORole) Delete() (error, error, int) {
